changed folder struct, refactored code
