+++ MyMessenger.Angular/src/app/layout/navbar/navbar.component.ts
@@ -1,5 +1,5 @@
 import { Component } from '@angular/core';
-import { AuthService } from '../../services/auth.service';
+import { AuthService } from '../../pages/auth/service/auth.service';
 import { CommonModule } from '@angular/common';
 
 @Component({
