+++ MyMessenger.Angular/src/app/pages/users/services/user.service.ts
@@ -1,9 +1,9 @@
 import { Injectable } from '@angular/core';
-import { DataGrid } from '../shared/models/datagrid';
-import { DataRetrieval } from '../shared/models/dataretrieval';
+import { DataGrid } from '../../../shared/interfaces/datagrid';
+import { DataRetrieval } from '../../../shared/interfaces/dataretrieval';
 import { HttpClient, HttpHeaders, HttpParams } from '@angular/common/http';
-import { Observable, map, of } from 'rxjs';
-import { environment } from '../../environments/environment';
+import { Observable, catchError, map, of } from 'rxjs';
+import { environment } from '../../../../environments/environment';
 
 @Injectable({
   providedIn: 'root'
@@ -16,19 +16,18 @@ export class UserService {
   constructor(private http: HttpClient) { }
 
   getAll<T>(data: DataRetrieval, endpoint: string): Observable<DataGrid<T>> {
-    try {
-      const headers = new HttpHeaders({
-        'Authorization': 'Bearer ' + this.accessToken
-      })
-      var params = this.setUrlParams(data);
-      return this.http.get<DataGrid<T>>(this.baseUrl + endpoint, { headers , params })
-      .pipe(
-        map(dataGrid => {
-        return dataGrid;
-      }));
-    } catch (error) {
-      return of({ data: [], numberOfPages: 1 });
-    }
+    const headers = new HttpHeaders({
+      'Authorization': 'Bearer ' + this.accessToken
+    })
+    var params = this.setUrlParams(data);
+    return this.http.get<DataGrid<T>>(this.baseUrl + endpoint, { headers , params })
+    .pipe(
+      map(dataGrid => {
+      return dataGrid;
+    }),
+    catchError(() => 
+      of({ data: [], numberOfPages: 1 }
+    )));
   }
   setUrlParams(data: DataRetrieval) : HttpParams {
     var params = new HttpParams()
