+++ MyMessenger.Angular/src/app/shared/interfaces/datagrid.ts
