+++ MyMessenger.Angular/src/app/shared/interfaces/dataretrieval.ts
